Fix command vector data types

+++ commandVector.h
@@ -12,10 +12,10 @@ struct CommandVector final {
 
     // bitfield order is {AUX1_low, AUX1_mid, AUX1_high, AUX2_low, AUX2_mid, AUX2_high, x, x} (LSB-->MSB)
     uint8_t aux_mask{0};
-    float throttle{0.0};
-    float pitch{0.0};
-    float roll{0.0};
-    float yaw{0.0};
+    int16_t throttle{0};
+    int16_t pitch{0};
+    int16_t roll{0};
+    int16_t yaw{0};
     Source source{Source::None};
 
     void clearBluetoothState() {
